Started code to re-do the analysis of the averaged waveforms

+++ analysis/qval_calculation_functions.h
@@ -1,31 +1,6 @@
-# include <cstdio>
-# include <cmath>
-# include <cstdlib>
-# include <iostream>
-# include <iomanip>
-# include <fstream>
-# include <sstream>
-# include <string>
-# include <cassert>
-# include <vector>
-# include <algorithm>
+#include "TProfile.h"
 
-# include "TCanvas.h"
-# include "TROOT.h"
-# include "TH1.h"
-# include "TH2.h"
-# include "TH1F.h"
-# include "TF1.h"
-# include "TApplication.h"
-# include "TMultiGraph.h"
-# include "TMath.h"
-# include "TTree.h"
-# include "TFile.h"
-# include "TLine.h"
-# include "TNtuple.h"
-# include "TProfile.h"
-
-void calculateQVal(int tailStart, int tailEnd, 
+void calculateQVal(double tailStart, double tailEnd, 
 				   TProfile* neutronWaveform, TProfile* photonWaveform,
 				   double &neutronQVal, double &photonQVal);
